requirement.txt
Django>=1.11,<2.0

Pillow
arrow
PyJWT
Unidecode
certifi
chardet
defusedxml
django-allauth
django-appconf
django-axes
django-ipware
django-static-sitemaps
django-simplemde
djangorestframework
djangorestframework-jwt
django-cors-headers
django-extensions
django-haystack
easy-thumbnails
python-memcached
user-agents
requests
requests-oauthlib
idna
oauthlib
python3-openid
six
